+++ routes/index.js
@@ -1,13 +1,32 @@
 const passport = require('passport');
 const route = require('express').Router();
-const { userLogin, loginPage } = require('../api/login');
+
+const jwt = require('../interceptors/jwt');
 const { verifyUserJWT } = require('../api/users');
+const { userLogin, loginPage } = require('../api/login');
 const { userRegistration, verifyUserRegistration } = require('../api/signup');
 
-route.post('/register', userRegistration);
-route.get('/register', verifyUserRegistration);
-route.post('/login', passport.authenticate('local'), userLogin);
+
+// Render Login Page (sample/test)
 route.get('/login', loginPage);
-route.post('/token', verifyUserJWT);
+
+// Login API
+route.post('/login', passport.authenticate('local'), userLogin);
+
+// Register New user
+route.post('/register', userRegistration);
+
+// Verify User and redirect to Login on success
+route.get('/register', verifyUserRegistration);
+
+// Verify JWT token and Fetch user details
+route.post('/token', jwt, verifyUserJWT);
+
+
+
+// route.post('/forgot-password', forgotPasswordOTP);
+// route.post('/forgot-password/otp', verifyForgotPasswordOTP);
+// route.put('/forgot-password/reset', resetNewPassword);
+// route.put('/password', jwt, updateNewPassword);
 
 module.exports = route;
